app/backend/requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
httpx
pydantic-settings
python-multipart
websockets
